+++ src/8thWonderland/langs/fr.php
@@ -77,7 +77,11 @@ $array = array(
     "motions_archives"      => "Motions archivées",
     "stats_country"         => "Statistiques du pays",
     'my_groups'             => 'Mes groupes',
-    "title_groups"          => "Groupes régionaux et Groupes thématiques",
+    
+    // Groupes
+    'groups'                => 'Groupes',
+    'groups.thematic'       => 'Groupes Thématiques',
+    'groups.regional'       => 'Groupes Régionaux',
         
 
     // Partie 'TACHES'